requirements.txt
langchain==0.3.13
langchain-openai==0.2.13
openai==1.55.3
python-dotenv==1.0.1
